# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: EDU Power Push Backend API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    API for campus electricity scanning, product user accounts, personal and
    campus statistics, push notifications, and the operator admin panel.
    Processes bind to loopback by default (Compose maps only 127.0.0.1). Public
    access is same-origin reverse proxy + TLS; operators may also use an SSH
    tunnel. Campus aggregates (`/api/v1/campus/*`) are anonymous; personal data
    uses short-lived access JWTs; admin and raw meter endpoints require an
    operator/admin session or `ADMIN_TOKEN` (never ship that token in a browser build).
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/me/refresh:
    post:
      tags: [Users]
      operationId: refreshCurrentUserMeter
      summary: Fetch and store one on-demand upstream reading for the bound meter
      description: >
        The server enforces a 30-second per-meter cooldown. Single-meter refreshes run in
        their own upstream lane, so a campus scan or monthly reconciliation no longer blocks
        them. status upstream_unavailable now means what it says: the upstream returned no
        usable reading, or the refresh lane was saturated, and the last stored reading is
        returned instead.
      security:
        - userAccessToken: []
      responses:
        "200":
          description: Refresh attempted; inspect status for refreshed, cached, or degraded outcome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeterRefresh"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: User has no active meter binding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/admin/scan-runs:
    get:
      tags: [Operations]
      operationId: listScanRuns
      summary: List scan runs
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ScanRunStatus"
      responses:
        "200":
          description: Scan run page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRunPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Operations]
      operationId: createScanRun
      summary: Start a controlled scan run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRunRequest"
      responses:
        "200":
          description: Dry-run scope validation result; no scan run is created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDryRun"
        "202":
          description: Scan run accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRun"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Another campus scan is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/meters/{meter}/overview:
    get:
      tags: [Meters]
      operationId: getMeterOverview
      summary: Get latest balance and recent consumption summary for a meter
      parameters:
        - $ref: "#/components/parameters/Meter"
      responses:
        "200":
          description: Meter overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeterOverview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Internal administrator token from ADMIN_TOKEN.
    userAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived product-user access JWT.
    refreshCookie:
      type: apiKey
      in: cookie
      name: edu_power_refresh
      description: Long-lived rotating refresh JWT stored as an HttpOnly SameSite cookie.
  parameters:
...
    Meter:
      name: meter
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid administrator token, access JWT, or refresh JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: A required dependency is not ready
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Request rate or resource-specific cooldown exceeded
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    CaptchaProvider:
      type: string
      enum: [disabled, turnstile]
...
    Decimal:
      type: string
      pattern: "^-?[0-9]+(\\.[0-9]+)?$"
      examples: ["43.87"]
    Availability:
      type: string
      enum: [ready, partial, insufficient_history, unavailable]
    Quality:
      type: object
      required: [eligible, covered, coverage_ratio, stale, anomalies]
      properties:
        eligible:
          type: integer
          minimum: 0
        covered:
          type: integer
          minimum: 0
        coverage_ratio:
          type: number
          minimum: 0
          maximum: 1
        stale:
          type: integer
          minimum: 0
        anomalies:
          type: integer
          minimum: 0
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        provider:
          $ref: "#/components/schemas/CaptchaProvider"
        challenge_url:
          type: string
          format: uri-reference
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    ScanRunStatus:
      type: string
      enum: [pending, running, completed, completed_with_errors, interrupted, canceled, failed]
...
    ScanCounters:
      type: object
      required:
        [inventory_total, excluded_total, eligible_total, processed_total, valid_total,
         stale_total, empty_total, error_total, parse_error_total, duplicate_reading_total]
      properties:
        inventory_total:
          type: integer
        excluded_total:
          type: integer
        eligible_total:
          type: integer
        processed_total:
          type: integer
        valid_total:
          type: integer
        stale_total:
          type: integer
        empty_total:
          type: integer
        error_total:
          type: integer
        parse_error_total:
          type: integer
        duplicate_reading_total:
          type: integer
    ScanRun:
      type: object
      required: [id, trigger, status, started_at, bound_only, counters]
      properties:
        id:
          type: string
          format: uuid
        parent_run_id:
          type: [string, "null"]
          format: uuid
        trigger:
          type: string
          enum: [manual, schedule, retry, recovery]
        status:
          $ref: "#/components/schemas/ScanRunStatus"
        started_at:
          type: string
          format: date-time
        finished_at:
          type: [string, "null"]
          format: date-time
        heartbeat_at:
          type: [string, "null"]
          format: date-time
        qps:
          type: number
        concurrency:
          type: integer
        limit:
          type: [integer, "null"]
        bound_only:
          type: boolean
          description: True for the hourly priority scan of meters bound to active users.
        counters:
          $ref: "#/components/schemas/ScanCounters"
    ScanRunPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanRun"
        next_cursor:
          type: [string, "null"]
    CreateScanRunRequest:
      type: object
      properties:
        limit:
          type: [integer, "null"]
          minimum: 1
        full:
          type: boolean
          default: false
          description: Must be explicitly true for an unbounded full-inventory scan.
        building:
          type: [string, "null"]
        floor:
          type: [string, "null"]
        qps:
          type: [number, "null"]
          minimum: 0.01
          maximum: 10
        concurrency:
          type: [integer, "null"]
          minimum: 1
          maximum: 16
        dry_run:
          type: boolean
          default: false
    ScanDryRun:
      type: object
      required: [dry_run, counters, qps, concurrency]
      properties:
        dry_run:
          type: boolean
          const: true
        counters:
          $ref: "#/components/schemas/ScanCounters"
        qps:
          type: number
        concurrency:
          type: integer
...
    LatestReading:
      type: object
      required: [reading_time, observed_at, prepaid_yuan, total_yuan, total_kwh, freshness]
      properties:
        reading_time:
          type: string
          format: date-time
        observed_at:
          type: string
          format: date-time
        prepaid_yuan:
          $ref: "#/components/schemas/Decimal"
        subsidy_yuan:
          oneOf:
            - $ref: "#/components/schemas/Decimal"
            - type: "null"
        total_yuan:
          $ref: "#/components/schemas/Decimal"
        total_kwh:
          $ref: "#/components/schemas/Decimal"
        meter_status:
          type: string
        freshness:
          type: string
          enum: [fresh, stale]
    MeterOverview:
      type: object
      required: [meter, location, availability, quality]
      properties:
        meter:
          type: string
        location:
          type: object
          required: [building, floor, room]
          properties:
            building:
              type: string
            floor:
              type: string
            room:
              type: string
        latest:
          oneOf:
            - $ref: "#/components/schemas/LatestReading"
            - type: "null"
        recent_7d_kwh:
          oneOf:
            - $ref: "#/components/schemas/Decimal"
            - type: "null"
        availability:
          $ref: "#/components/schemas/Availability"
        quality:
          $ref: "#/components/schemas/Quality"
    MeterRefresh:
      type: object
      required: [meter, status, latest, refreshed_at, next_allowed_at, availability, quality]
      properties:
        meter:
          type: string
        status:
          type: string
          enum: [refreshed, cached, upstream_unavailable]
        latest:
          oneOf:
            - $ref: "#/components/schemas/LatestReading"
            - type: "null"
        refreshed_at:
          type: string
          format: date-time
        next_allowed_at:
          type: string
          format: date-time
        availability:
          $ref: "#/components/schemas/Availability"
        quality:
          $ref: "#/components/schemas/Quality"
...